requirements.txt
azure-core>=1.29.6
azure-identity>=1.15.0
click>=8.1.7
colorama>=0.4.6
pandas>=2.2.2
requests>=2.32.3
rich>=13.7.1
